adapt gui
