+++ component/io/viz_io.py
@@ -3,10 +3,12 @@ from component import parameter as cp
 class VizIo:
     
     def __init__(self):
-        #input
+        # inputs
         self.check = False
         self.sources = None
         self.bands = None
         self.start_year = 2005
-        self.end_year = cp.max_end_year
+        self.end_year = cp.gee_max_end_year
         self.square_size = 2000 #in meters
+        self.planet_key = None
+        self.driver = None
