+++ component/parameter/__init__.py
@@ -1,4 +1,5 @@
 from .directory import *
 from .gee import *
 from .pdf import *
-from .satellites import *
+from .planet import *
+from .viz import *
