+++ component/parameter/planet.py
@@ -1,0 +1,20 @@
+
+
+# set the years that will be used with planet data
+planet_min_start_year = 2016
+planet_max_end_year = 2021
+
+# to each year select a date range that will match a basempa included in the SEPAL NICFI contract
+planet_date_ranges = {
+    2016: '2015-12_2016-05',
+    2017: '2016-12_2017-05',
+    2018: '2017-12_2018-05',
+    2019: '2018-12_2019-05',
+    2020: '2019-12_2012-05',
+    2021: '2021-01'
+}
+
+planet_bands_combo = [
+    "rgb",
+    "cir"
+]
