+++ component/parameter/viz.py
@@ -1,0 +1,6 @@
+# the drivers that can be used to create thumbnails
+drivers = ['gee', 'planet']
+
+# define the extrems size of the thumbnails
+min_square = 500
+max_square = 10000
